wechat-scrapy/requirements.txt
APScheduler==3.5.3
astroid==2.1.0
autopep8==1.4.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.1
DateTime==4.3
flake8==3.6.0
Flask==1.0.2
future==0.17.1
idna==2.7
isort==4.3.4
itchat==1.2.32
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pipenv==2018.11.26
pycodestyle==2.4.0
pycrypto==2.6.1
pyflakes==2.0.0
pylint==2.2.0
pymongo==3.7.2
pypng==0.0.18
PyQRCode==1.2.1
pytz==2018.7
requests==2.20.1
schedule==0.5.0
six==1.11.0
tzlocal==1.5.1
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
Werkzeug==0.14.1
wrapt==1.10.11
wxpy==0.3.9.8
yapf==0.25.0
zope.interface==4.6.0
